# Conflicts:
#	CMakeLists.txt
#	docs/tutorial.rst
#	src/liblsquic/lsquic_stream.h

+++ src/liblsquic/lsquic_parse_gquic_be.h
@@ -2,6 +2,8 @@
 #ifndef LSQUIC_PARSE_GQUIC_BE_H
 #define LSQUIC_PARSE_GQUIC_BE_H
 
+#include <stdint.h>
+
 /* Header file to make it easy to reference gQUIC parsing functions.  This
  * is only meant to be used internally.  The alternative would be to place
  * all gQUIC-big-endian functions -- from all versions -- in a single file,
@@ -9,7 +11,7 @@
  */
 
 #define CHECK_SPACE(need, pstart, pend)  \
-    do { if ((intptr_t) (need) > ((pend) - (pstart))) { return -1; } } while (0)
+    do { if ((uint64_t) (need) > (uint64_t) ((pend) - (pstart))) { return -1; } } while (0)
 
 uint64_t
 lsquic_gquic_be_read_float_time16 (const void *mem);
